feat(multi channel support): code cleanup, add comments
