+++ src/Kentico.Xperience.Lucene.Admin/Components/LuceneIndexConfigurationComponent.cs
@@ -2,7 +2,6 @@
 using CMS.DataEngine;
 
 using Kentico.Xperience.Admin.Base;
-using Kentico.Xperience.Admin.Base.FormAnnotations;
 using Kentico.Xperience.Admin.Base.Forms;
 using Kentico.Xperience.Lucene.Admin;
 using Kentico.Xperience.Lucene.Core.Indexing;
@@ -14,34 +13,38 @@ using Kentico.Xperience.Lucene.Core.Indexing;
 
 namespace Kentico.Xperience.Lucene.Admin;
 
-#pragma warning disable S2094 // intentionally empty class
-public class LuceneIndexConfigurationComponentProperties : FormComponentProperties
-{
-}
-#pragma warning restore
-
-public class LuceneIndexConfigurationComponentClientProperties : FormComponentClientProperties<IEnumerable<LuceneIndexChannelConfiguration>>
-{
-    public IEnumerable<LuceneIndexContentType>? PossibleContentTypeItems { get; set; }
-    public IEnumerable<LuceneIndexChannel>? PossibleChannels { get; set; }
-}
-
-public sealed class LuceneIndexConfigurationComponentAttribute : FormComponentAttribute
-{
-}
-
+/// <summary>
+/// Represents a form component for configuring Lucene index channels and their associated properties.
+/// </summary>
 [ComponentAttribute(typeof(LuceneIndexConfigurationComponentAttribute))]
 public class LuceneIndexConfigurationComponent : FormComponent<LuceneIndexConfigurationComponentProperties, LuceneIndexConfigurationComponentClientProperties, IEnumerable<LuceneIndexChannelConfiguration>>
 {
+    /// <summary>
+    /// Represents the unique identifier for the Lucene index configuration in the Kentico Xperience integrations.
+    /// </summary>
     public const string IDENTIFIER = "kentico.xperience-integrations-lucene-admin.lucene-index-configuration";
+
 
     internal List<LuceneIndexChannelConfiguration>? Value { get; set; }
 
+
+    /// <inheritdoc/>
     public override string ClientComponentName => "@kentico/xperience-integrations-lucene-admin/LuceneIndexConfiguration";
 
+
+    /// <inheritdoc/>
     public override IEnumerable<LuceneIndexChannelConfiguration> GetValue() => Value ?? [];
+
+
+    /// <inheritdoc/>
     public override void SetValue(IEnumerable<LuceneIndexChannelConfiguration> value) => Value = value.ToList();
 
+
+    /// <summary>
+    /// Deletes the <see cref="LuceneIndexChannelConfiguration"/> with the specified channel name.
+    /// </summary>
+    /// <param name="channelName">The name of the channel to be removed.</param>
+    /// <returns>The <see cref="ICommandResponse{RowActionResult}"/>.</returns>
     [FormComponentCommand]
     public Task<ICommandResponse<RowActionResult>> DeleteWebsiteChannelConfiguration(string channelName)
     {
@@ -54,6 +57,12 @@ public class LuceneIndexConfigurationComponent : FormComponent<LuceneIndexConfig
         return Task.FromResult(ResponseFrom(new RowActionResult(false)));
     }
 
+
+    /// <summary>
+    /// Saves the specified website channel configuration by updating or adding it to the current collection.
+    /// </summary>
+    /// <param name="channelConfiguration">The <see cref="LuceneIndexChannelConfiguration"/> object representing the website channel configuration to save.</param>
+    /// <returns>The <see cref="ICommandResponse{RowActionResult}"/>.</returns>
     [FormComponentCommand]
     public Task<ICommandResponse<RowActionResult>> SaveWebsiteChannelConfiguration(LuceneIndexChannelConfiguration channelConfiguration)
     {
@@ -69,6 +78,12 @@ public class LuceneIndexConfigurationComponent : FormComponent<LuceneIndexConfig
         return Task.FromResult(ResponseFrom(new RowActionResult(false)));
     }
 
+
+    /// <summary>
+    /// Adds a website channel configuration to the current collection if it does not already exist.
+    /// </summary>
+    /// <param name="websiteChannelName">The name of the website channel to add.</param>
+    /// <returns>The <see cref="ICommandResponse{RowActionResult}"/>.</returns>
     [FormComponentCommand]
     public Task<ICommandResponse<RowActionResult>> AddWebsiteChannelConfiguration(string websiteChannelName)
     {
@@ -88,6 +103,8 @@ public class LuceneIndexConfigurationComponent : FormComponent<LuceneIndexConfig
         }
     }
 
+
+    /// <inheritdoc/>
     protected override async Task ConfigureClientProperties(LuceneIndexConfigurationComponentClientProperties properties)
     {
         var allWebsiteChannels = (await ChannelInfoProvider.ProviderObject
