+++ src/Kentico.Xperience.Lucene.Admin/Components/LuceneIndexConfigurationComponentAttribute.cs
@@ -1,0 +1,10 @@
+﻿using Kentico.Xperience.Admin.Base.FormAnnotations;
+
+namespace Kentico.Xperience.Lucene.Admin;
+
+/// <summary>
+/// Represents an attribute used to mark a form component as a Lucene index configuration component.
+/// </summary>
+public sealed class LuceneIndexConfigurationComponentAttribute : FormComponentAttribute
+{
+}
