+++ src/Kentico.Xperience.Lucene.Admin/Components/LuceneIndexConfigurationComponentProperties.cs
@@ -1,0 +1,15 @@
+﻿using Kentico.Xperience.Admin.Base.Forms;
+
+namespace Kentico.Xperience.Lucene.Admin;
+
+#pragma warning disable IDE0079 // Remove unnecessary suppression
+#pragma warning disable S2094 // intentionally empty class
+
+/// <summary>
+/// Represents the configuration properties for a Lucene index component.
+/// </summary>
+public class LuceneIndexConfigurationComponentProperties : FormComponentProperties
+#pragma warning restore IDE0079 // Remove unnecessary suppression
+{
+}
+#pragma warning restore
